requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
base58==2.1.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.2.1
defusedxml==0.7.1
did-resolver==0.0.3
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.11.2
djangorestframework==3.16.0
frozenlist==1.7.0
grpcio==1.68.1
grpcio-tools==1.68.1
gunicorn==23.0.0
h11==0.16.0
hedera-sdk-py==2.50.0
hiero-sdk-python==0.1.1
hiero_did_sdk_python==0.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.7
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.6.4
multipledispatch==1.0.0
nh3==0.2.21
oauthlib==3.2.2
packaging==25.0
pillow==10.4.0
propcache==0.3.2
protobuf==5.28.1
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyjnius==1.7.0
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytube==15.0.0
pytz==2025.2
qrcode==8.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
toolz==0.12.1
tqdm==4.66.5
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
unification==0.2.2
urllib3==2.2.2
uvicorn==0.35.0
whitenoise==6.7.0
xyzservices==2024.6.0
yarl==1.20.1
zope.interface==7.0.3
zstandard==0.23.0
